% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_marker_channel.R
\name{clean_marker_channel}
\alias{clean_marker_channel}
\title{Clean the marker channel of an fEMG recording}
\usage{
clean_marker_channel(
  femg_data,
  marker_channel,
  valid_marker_values,
  invalid_marker_values = c(),
  denoise_width = "auto"
)
}
\arguments{
\item{femg_data}{data frame or tibble}

\item{marker_channel}{the name of the column in femg_data that corresponds to the marker channel}

\item{valid_marker_values}{valid marker values to be found in marker_channel that are used to indicate the stimulus/task}

\item{invalid_marker_values}{invalid marker values that can safely be set to 0/off}

\item{denoise_width}{an integer or "auto". The number of samples for which to take the median as a method to remove individual samples with unexpected values.}
}
\value{
femg_data with additional column (marker_channel)_cleaned
}
\description{
Clean the marker channel of an fEMG recording
}
\examples{
clean_marker_channel(
    femg_raw_data_sample,
    marker_channel = "Marker",
    valid_marker_values = c(54)
    )
}
