% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countenance-package.R
\docType{package}
\name{countenance-package}
\alias{countenance}
\alias{countenance-package}
\title{countenance: Read, process and analyse facial EMG data}
\description{
Read, process and analyse facial EMG data. This has been written and tested for data collected with BIOPAC AcqKnowledge and exported as text, but can be used with any data in a similar format.
}
\author{
\strong{Maintainer}: Sarah McIntyre \email{sarah.mcintyre@liu.se} (\href{https://orcid.org/0000-0002-0544-6533}{ORCID}) (@somacrat)

}
\keyword{internal}
