% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countenance-package.R
\docType{package}
\name{countenance}
\alias{countenance}
\alias{countenance-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/countenance}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/countenance/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
