% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countenance_example.R
\name{countenance_example}
\alias{countenance_example}
\title{Get path to countenance example data}
\usage{
countenance_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If NULL, the example files will be listed.}
}
\value{
path and file name
}
\description{
Get path to countenance example data
}
\examples{
countenance_example("femg_raw_exported_sample.txt")
}
