% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_marker_channel.R
\name{denoise}
\alias{denoise}
\title{Denoise}
\usage{
denoise(marker_channel, denoise_width = "auto", values_to_ignore = c())
}
\arguments{
\item{marker_channel}{a vector, the marker channel to denoise}

\item{denoise_width}{an integer or "auto". The number of samples for which to take the median as a method to remove individual samples with unexpected values.}

\item{values_to_ignore}{a vector of values in marker_channel that should not be subject to denoising}
}
\value{
a list of flagged indices, and the full cleaned marker channel
}
\description{
Internal function called by clean_marker_channel().
}
