% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_marker_values.R
\name{fill_marker_values}
\alias{fill_marker_values}
\title{Title}
\usage{
fill_marker_values(
  femg_data,
  fill_duration,
  marker_channel,
  time_channel = "time_sec"
)
}
\arguments{
\item{femg_data}{data frame or tibble}

\item{fill_duration}{duration to fill in the marker channel for each marker}

\item{marker_channel}{the name of the column in femg_data that corresponds to the marker channel}

\item{time_channel}{the name of the column in femg_data that gives the time in seconds}
}
\value{

}
\description{
Title
}
