% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{pi_fwd_dest}
\alias{pi_fwd_dest}
\title{Get forward destinations (i.e. upstream DNS)}
\usage{
pi_fwd_dest(host = pi_host(), api_key = pi_key())
}
\arguments{
\item{host}{See \code{\link[=pi_host]{pi_host()}}}

\item{api_key}{See \code{\link[=pi_key]{pi_key()}}}
}
\description{
Get forward destinations (i.e. upstream DNS)
}
