% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi-host.R
\name{pi_host}
\alias{pi_host}
\title{Get or set PIHOLE_HOST value}
\usage{
pi_host(force = FALSE)
}
\arguments{
\item{force}{Force setting a new Pi-Hole host\link{:port} the current environment?}
}
\value{
atomic character vector containing the Pi-Hole host\link{:port}
}
\description{
Get or set PIHOLE_HOST value
}
