% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overtime-10.R
\name{pi_over_time_10m}
\alias{pi_over_time_10m}
\title{Get recent 10 min time series counts}
\usage{
pi_over_time_10m(host = pi_host(), api_key = pi_key())
}
\arguments{
\item{host}{See \code{\link[=pi_host]{pi_host()}}}

\item{api_key}{See \code{\link[=pi_key]{pi_key()}}}
}
\description{
Get recent 10 min time series counts
}
