% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-q.R
\name{pi_queries}
\alias{pi_queries}
\title{Get query types distribution}
\usage{
pi_queries(q = NULL, host = pi_host(), api_key = pi_key())
}
\arguments{
\item{q}{query}

\item{host}{See \code{\link[=pi_host]{pi_host()}}}

\item{api_key}{See \code{\link[=pi_key]{pi_key()}}}
}
\description{
Get query types distribution
}
