% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-types.R
\name{pi_query_types}
\alias{pi_query_types}
\title{Get query types distribution}
\usage{
pi_query_types(host = pi_host(), api_key = pi_key())
}
\arguments{
\item{host}{See \code{\link[=pi_host]{pi_host()}}}

\item{api_key}{See \code{\link[=pi_key]{pi_key()}}}
}
\description{
Get query types distribution
}
