% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent.R
\name{pi_recent}
\alias{pi_recent}
\title{Get recent blocked}
\usage{
pi_recent(host = pi_host(), api_key = pi_key())
}
\arguments{
\item{host}{See \code{\link[=pi_host]{pi_host()}}}

\item{api_key}{See \code{\link[=pi_key]{pi_key()}}}
}
\description{
Get recent blocked
}
