% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{pi_summary}
\alias{pi_summary}
\title{Get summary stats}
\usage{
pi_summary(host = pi_host(), api_key = pi_key())
}
\arguments{
\item{host}{See \code{\link[=pi_host]{pi_host()}}}

\item{api_key}{See \code{\link[=pi_key]{pi_key()}}}
}
\description{
Get summary stats
}
