% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pi_test}
\alias{pi_test}
\alias{pi_type}
\alias{pi_version}
\alias{pi_enable}
\alias{pi_disable}
\title{Test connectivity to/get info about a Pi-Hole}
\usage{
pi_test(host = pi_host(), api_key = pi_key())

pi_type(host = pi_host(), api_key = pi_key())

pi_version(host = pi_host(), api_key = pi_key())

pi_enable(host = pi_host(), api_key = pi_key())

pi_disable(host = pi_host(), api_key = pi_key())
}
\arguments{
\item{host}{See \code{\link[=pi_host]{pi_host()}}}

\item{api_key}{See \code{\link[=pi_key]{pi_key()}}}
}
\description{
Test connectivity to/get info about a Pi-Hole
}
