% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-clients.R
\name{pi_top_clients}
\alias{pi_top_clients}
\title{Get top clients}
\usage{
pi_top_clients(n = 100, host = pi_host(), api_key = pi_key())
}
\arguments{
\item{n}{\# of top clients (default 100)}

\item{host}{See \code{\link[=pi_host]{pi_host()}}}

\item{api_key}{See \code{\link[=pi_key]{pi_key()}}}
}
\description{
Get top clients
}
