% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-items.R
\name{pi_top_items}
\alias{pi_top_items}
\title{Get top items}
\usage{
pi_top_items(n = 100L, host = pi_host(), api_key = pi_key())
}
\arguments{
\item{n}{\# of top items (default 100)}

\item{host}{See \code{\link[=pi_host]{pi_host()}}}

\item{api_key}{See \code{\link[=pi_key]{pi_key()}}}
}
\description{
Get top items
}
