% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marker_sequence.R
\name{plot_marker_sequence}
\alias{plot_marker_sequence}
\title{Plot the sequence of marker values}
\usage{
plot_marker_sequence(
  femg_data,
  marker_channel,
  unexpected_channel = "unexpected",
  time_channel = "time_sec"
)
}
\arguments{
\item{femg_data}{data frame or tibble}

\item{marker_channel}{the name of the column in femg_data that corresponds to the marker channel}

\item{unexpected_channel}{the name of the column in femg_data that is a boolean indicating if the channel was unexpected}

\item{time_channel}{the name of the column in femg_data that gives the time in seconds}
}
\value{
a ggplot object
}
\description{
Create a ggplot object, plotting the sequence of transitions found in a marker channel, and highlighting unexpected values in red.
}
