% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acq_text.R
\name{read_acq_text}
\alias{read_acq_text}
\title{Read a text file exported from BIOPAC AcqKnowledge}
\usage{
read_acq_text(file_name, delim, keep_channels)
}
\arguments{
\item{file_name}{Name of data file to read.}

\item{delim}{Single character used to separate fields within a record.}

\item{keep_channels}{A vector with the names of the channels in the data file to keep. All others are discarded.}
}
\value{
A tibble.
}
\description{
Read a text file exported from BIOPAC AcqKnowledge
}
\examples{
# read_acq_text(
#     "folder/file.txt",
#     delim = ",",
#     keep_channels = c("Zyg","Corr","Marker")
#     )
}
