% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countenance-package.R
\docType{package}
\name{countenance}
\alias{countenance}
\alias{countenance-package}
\title{Tools to Work with the Pi-Hole API}
\description{
Named after a primary synonym for 'pihole', tools are provided
to access the Pi-Hole API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/countenance}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/countenance/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
