% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acq_text.R
\name{read_acq_text}
\alias{read_acq_text}
\title{Read a text file exported from BIOPAC AcqKnowledge.}
\usage{
read_acq_text(file, delim, keep_channels)
}
\arguments{
\item{file}{Name of data file to read.}

\item{delim}{Single character used to separate fields within a record.}

\item{keep_channels}{A vector with the names of the channels in the data file to keep. All others are discarded.}
}
\value{
A tibble.
}
\description{
Wrapper around readr::read_delim(). Additionally processes header lines with readr::read_lines(), and adds a time variable for convenient plotting.
}
\examples{
read_acq_text(
    countenance_example("femg_raw_exported_sample.txt"),
    delim = ",",
    keep_channels = c("Zyg Processed","Corr Processed","Marker")
    )
}
